:sparkles: Run task on user subscription date update

+++ telegram_bot/admin.py
@@ -2,9 +2,15 @@ import logging
 
 from django.contrib import admin
 
-from telegram_bot.models import User
+from telegram_bot.models import User, ScheduledTask
 from wireguard.exceptions import VPNServerError
 from wireguard.services.vpn_server import VPNServerService
+from wireguard.tasks import on_user_subscription_date_updated
+
+
+@admin.register(ScheduledTask)
+class ScheduledTaskAdmin(admin.ModelAdmin):
+    pass
 
 
 class UserInline(admin.TabularInline):
@@ -43,22 +49,25 @@ class UserAdmin(admin.ModelAdmin):
 
     def save_model(self, request, obj, form, change):
         old_user: User = self.model.objects.filter(id=obj.id).select_related('server').first()
-        if old_user is not None:
+        if old_user is not None and obj.subscribed_at != old_user.subscribed_at:
+            on_user_subscription_date_updated.delay(telegram_id=obj.telegram_id)
+        if obj.server.url != old_user.server.url:
+            if old_user is not None:
+                try:
+                    with VPNServerService(base_url=old_user.server.url,
+                                          password=old_user.server.password) as old_vpn_server:
+                        old_vpn_server.login()
+                        old_vpn_server.delete_user(old_user.uuid)
+                except VPNServerError:
+                    logging.error('Could not delete user from old server')
             try:
-                with VPNServerService(base_url=old_user.server.url,
-                                      password=old_user.server.password) as old_vpn_server:
-                    old_vpn_server.login()
-                    old_vpn_server.delete_user(old_user.uuid)
+                with VPNServerService(base_url=obj.server.url, password=obj.server.password) as new_vpn_server:
+                    new_vpn_server.login()
+                    user_in_vpn_server, _ = new_vpn_server.get_or_create_user(obj.telegram_id)
             except VPNServerError:
-                logging.error('Could not delete user from old server')
-        try:
-            with VPNServerService(base_url=obj.server.url, password=obj.server.password) as new_vpn_server:
-                new_vpn_server.login()
-                user_in_vpn_server, _ = new_vpn_server.get_or_create_user(obj.telegram_id)
-        except VPNServerError:
-            logging.error('Could not create user in new server')
-        else:
-            obj.uuid = user_in_vpn_server.uuid
+                logging.error('Could not create user in new server')
+            else:
+                obj.uuid = user_in_vpn_server.uuid
         super().save_model(request, obj, form, change)
 
     def delete_model(self, request, obj):
